+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- NFS export failed to set correct address type in IPv6 clusters.
+
 ## [1.10.5] - 2025-12-16
 
 ### Changed